feat: port database from Postgres to SQLite
- internal/db/db.go on modernc.org/sqlite, WAL + busy_timeout pragmas,
  SetMaxOpenConns(1)
- goose dialect "sqlite3"; squash 10 migrations into a single
  00001_initial.sql; data-only backfills dropped (window-function DELETE,
  ::jsonb backfill, archived-rescue)
- sqlc engine "sqlite"; queries moved postgres/ -> sqlite/; $N -> ?;
  ::text casts removed; IS NOT DISTINCT FROM -> IS;
  em-dashes in comments -> ASCII (caused sqlc-sqlite "SELECid" cascade
  parse errors via UTF-8 offset miscount); now() -> strftime
- internal/core/activity/list.go: ILIKE -> LIKE; search term bound twice
  because positional ? can't reuse args
- DatabaseConfig collapses to {Path}; removed DefaultDBDriver and dead
  sqlite branches in load.go, router.go, system.go doc
- internal/core/dbutil: passthrough Null helpers + NullableString +
  IsUniqueViolation against *modernc/sqlite.Error
- 3 Dockerfiles: PILOT_DATABASE_PATH=/config/pilot.db; drop pgx ENV
- Tests: pgx-based mock unique-violation rebuilt via real in-memory
  SQLite trigger; load_test password-file cases removed
- int32 -> int64 / sql.NullString -> *string / TIMESTAMPTZ -> TEXT
  ripple across ~90 sites in ~14 packages

+++ internal/core/stats/service.go
@@ -274,11 +274,11 @@ func (s *Service) Snapshot(ctx context.Context) error {
 
 	return s.q.InsertStatsSnapshot(ctx, db.InsertStatsSnapshotParams{
 		ID:                uuid.New().String(),
-		TotalSeries:       int32(col.TotalSeries),
-		TotalEpisodes:     int32(col.TotalEpisodes),
-		MonitoredEpisodes: int32(col.Monitored),
-		WithFile:          int32(col.WithFile),
-		Missing:           int32(col.Missing),
+		TotalSeries:       col.TotalSeries,
+		TotalEpisodes:     col.TotalEpisodes,
+		MonitoredEpisodes: col.Monitored,
+		WithFile:          col.WithFile,
+		Missing:           col.Missing,
 		TotalSizeBytes:    stor.TotalBytes,
 		SnapshotAt:        time.Now().UTC().Format(time.RFC3339),
 	})
